+++ packages/compiler/test/RuntimeSliceNative.test.ts
@@ -6,7 +6,7 @@ import * as Analysis from '../src/Analysis.js'
 const ascii = (value: string): Uint8Array =>
   Uint8Array.from(value, (character) => character.charCodeAt(0))
 
-const source = `import silk.usize as usize
+const source = `import silk.usize
 struct Pair { left: i32 right: i32 }
 fn replace(values: &mut [Pair], index: usize) -> i32 {
   values[index] = Pair { left: 40, right: 2 }
