+++ packages/compiler/test/StringOwnership.test.ts
@@ -11,7 +11,7 @@ const snapshot = (source: string) => Analysis.ofSource('string/ownership', encod
 
 it.effect('retains a runtime string backing loan through its last use', () =>
   Effect.gen(function* () {
-    const source = `import silk.u8 as u8
+    const source = `import silk.u8
 fn conflict() -> usize {
   let mut bytes = [u8.toU8(104), u8.toU8(195), u8.toU8(169)]
   unsafe {
@@ -80,7 +80,7 @@ pub fn main() -> i32 { return 0 }`
 
 it.effect('propagates string loans through returned views and ordinary calls', () =>
   Effect.gen(function* () {
-    const source = `import silk.u8 as u8
+    const source = `import silk.u8
 fn view(bytes: &[u8]) -> string {
   unsafe { return Intrinsic.stringFromUtf8Unchecked(bytes) }
 }
@@ -120,7 +120,7 @@ pub fn main() -> i32 { return 0 }`
 
 it.effect('preserves loans nested inside generic result data', () =>
   Effect.gen(function* () {
-    const source = `import silk.u8 as u8
+    const source = `import silk.u8
 import silk.result { Result }
 struct InvalidUtf8 { offset: usize }
 fn validate(bytes: &[u8], accepted: bool) -> Result<string, InvalidUtf8> {
