+++ packages/compiler/test/support/floatMath.ts
@@ -16,7 +16,7 @@ export interface FloatMathProgram {
   readonly source: string
 }
 
-const f64Exact = `import silk.f64 as f64
+const f64Exact = `import silk.f64
 pub fn main() -> i32 {
   // Negative zero survives every operation in the exact group.
   if f64.toBits(f64.abs(-0.0)) != f64.toBits(0.0) { return 1 }
@@ -67,7 +67,7 @@ pub fn main() -> i32 {
   return 42
 }`
 
-const f32Exact = `import silk.f32 as f32
+const f32Exact = `import silk.f32
 pub fn main() -> i32 {
   if f32.toBits(f32.abs(-0.0)) != f32.toBits(0.0) { return 1 }
   if f32.toBits(f32.floor(-0.0)) != f32.toBits(-0.0) { return 2 }
@@ -97,8 +97,8 @@ pub fn main() -> i32 {
   return 42
 }`
 
-const minimumMaximum = `import silk.f32 as f32
-import silk.f64 as f64
+const minimumMaximum = `import silk.f32
+import silk.f64
 pub fn main() -> i32 {
   if f64.min(1.0, 2.0) != 1.0 { return 1 }
   if f64.max(1.0, 2.0) != 2.0 { return 2 }
@@ -128,8 +128,8 @@ pub fn main() -> i32 {
  * independently of the compiler, from a host square root, which IEEE-754 requires be correctly
  * rounded and which therefore agrees with any conforming implementation.
  */
-const squareRoot = `import silk.f32 as f32
-import silk.f64 as f64
+const squareRoot = `import silk.f32
+import silk.f64
 pub fn main() -> i32 {
   // sqrt(2), sqrt(3), sqrt(10) — irrational roots that pin the rounding of the last bit.
   if f64.toBits(f64.sqrt(f64.fromBits(4611686018427387904))) != 4609047870845172685 { return 1 }
@@ -172,8 +172,8 @@ pub fn main() -> i32 {
  * Requirement 10: every function reports the one canonical NaN encoding for a NaN input, so no
  * engine's payload propagation rule can show through.
  */
-const canonicalNaN = `import silk.f32 as f32
-import silk.f64 as f64
+const canonicalNaN = `import silk.f32
+import silk.f64
 fn notANumber(zero: f64) -> f64 { return zero / zero }
 fn notANumber32(zero: f32) -> f32 { return zero / zero }
 pub fn main() -> i32 {
